First commit.
Survival Games X is finally open source as we prepare for a complete
re-write as the team focuses on heavy optimization, a cleaner framework,
polished and streamlined code, more events & efficient API for
developers, better configuration, better MySQL & SQLite support, better
rollback, MORE features for you, and much much more behind the scenes
work! All of this is necessary and will make it much easier for the team
to update Survival Games X in the near future.

Please remember that, at this moment, you are contributing to the OLD
source code.

Contributing guidelines & the Survival Games X wiki will be coming soon!

We are excited for the future of Survival Games X! Thanks for the
support over this past year.